Move extractors to a package
