+++ extract/Extractor.py
