+++ tests/test_extractor.py
@@ -12,9 +12,9 @@ from pyspark.sql.types import (
 
 from lib.Utils import get_spark_session
 
-from lib.CsvExtractor import CsvExtractor
-from lib.HiveExtractor import HiveExtractor
-from lib.ExtractorFactoryImpl import ExtractorFactoryImpl
+from extract.CsvExtractor import CsvExtractor
+from extract.HiveExtractor import HiveExtractor
+from extract.ExtractorFactoryImpl import ExtractorFactoryImpl
 from lib.MissingSchemaError import MissingSchemaError
 from lib.MissingSourceLocationError import MissingSourceLocationError
